requirements.txt
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.6.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==2.0
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
WTForms==2.2.1
